CLN: refactor reader classes in shg_pl_trans module
